Release datarouter 0.0.119

+++ datarouter-bytes/src/main/java/io/datarouter/bytes/VarIntTool.java
@@ -248,7 +248,7 @@ public class VarIntTool{
 		return downcastToInt(decodeLong(bytes, 0));
 	}
 
-	public static long decodeInt(InputStream inputStream){
+	public static int decodeInt(InputStream inputStream){
 		return downcastToInt(decodeLongInternal(inputStream));
 	}
